# docs/swagger.yaml
basePath: /api/v1
definitions:
...
  meal.CreateMealDto:
    properties:
      calories:
        type: number
      image:
        type: string
      ingredients:
        items:
          $ref: '#/definitions/types.Ingredient'
        type: array
      name:
        type: string
      nutrients:
        items:
          $ref: '#/definitions/types.Nutrient'
        type: array
      userid:
        type: string
    required:
    - name
    type: object
  meal.Meal:
    properties:
      calories:
        default: 0
        type: number
      created_at:
        default: "null"
        type: string
      id:
        type: string
      image:
        default: "null"
        type: string
      ingredients:
        items:
          $ref: '#/definitions/types.Ingredient'
        type: array
      name:
        type: string
      nutrients:
        items:
          $ref: '#/definitions/types.Nutrient'
        type: array
      updated_at:
        default: "null"
        type: string
      userid:
        type: string
    required:
    - name
    type: object
  meal.UpdateMealDto:
    properties:
      calories:
        type: number
      carb:
        type: number
      fat:
        type: number
      protein:
        type: number
    type: object
  plans.CreatePlanDto:
    properties:
      dayofweek:
        type: string
      exercise:
        items:
          type: string
        type: array
      typeofplan:
        default: Rest
        type: string
      userid:
        type: string
    required:
    - dayofweek
    - userid
    type: object
  plans.Plan:
    properties:
      dayofweek:
        type: string
      exercise:
        items:
          type: string
        type: array
      id:
        type: string
      typeofplan:
        type: string
      userid:
        type: string
    required:
    - dayofweek
    - typeofplan
    - userid
    type: object
  plans.UpdatePlanDto:
    properties:
      exercise:
        items:
          type: string
        type: array
      typeofplan:
        type: string
    type: object
  types.Ingredient:
    properties:
      ingredientid:
        type: string
      numofservings:
        type: number
      unit:
        type: string
      value:
        type: number
    required:
    - ingredientid
    - numofservings
    - unit
    - value
    type: object
  types.Nutrient:
    properties:
      category:
        type: string
      name:
        type: string
      type:
        type: string
      unit:
        type: string
      value:
        type: number
    required:
    - name
    - type
    - unit
    - value
    type: object
...
host: localhost:8080
info:
  contact: {}
  description: This is a sample server for GymsBro API.
  title: GymsBro API
  version: "1.0"
paths:
...
  /meals:
    get:
      consumes:
      - application/json
      description: Get all meals
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/meal.Meal'
            type: array
        "400":
          description: Bad Request
          schema: {}
      summary: Get all meals
      tags:
      - meals
    post:
      consumes:
      - application/json
      description: Create a meal
      parameters:
      - description: Create Meal
        in: body
        name: meal
        required: true
        schema:
          $ref: '#/definitions/meal.CreateMealDto'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/meal.Meal'
        "400":
          description: Bad Request
          schema: {}
      summary: Create a meal
      tags:
      - meals
  /meals/{id}:
    delete:
      consumes:
      - application/json
      description: Delete a meal
      parameters:
      - description: Meal ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema: {}
      summary: Delete a meal
      tags:
      - meals
    get:
      consumes:
      - application/json
      description: Get a meal
      parameters:
      - description: Meal ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/meal.Meal'
        "400":
          description: Bad Request
          schema: {}
      summary: Get a meal
      tags:
      - meals
    put:
      consumes:
      - application/json
      description: Update a meal
      parameters:
      - description: Meal ID
        in: path
        name: id
        required: true
        type: string
      - description: Update Meal
        in: body
        name: meal
        required: true
        schema:
          $ref: '#/definitions/meal.UpdateMealDto'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/meal.Meal'
        "400":
          description: Bad Request
          schema: {}
      summary: Update a meal
      tags:
      - meals
  /meals/user/{userid}:
    get:
      consumes:
      - application/json
      description: Get meals by user
      parameters:
      - description: User ID
        in: path
        name: userid
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/meal.Meal'
            type: array
        "400":
          description: Bad Request
          schema: {}
      summary: Get meals by user
      tags:
      - meals
  /meals/userdate/{userid}:
    get:
      consumes:
      - application/json
      description: Get meals by user and date
      parameters:
      - description: User ID
        in: path
        name: userid
        required: true
        type: string
      - description: Start date
        in: query
        name: start
        required: true
        type: integer
      - description: End date
        in: query
        name: end
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/meal.Meal'
            type: array
        "400":
          description: Bad Request
          schema: {}
      summary: Get meals by user and date
      tags:
      - meals
  /plans:
    get:
      consumes:
      - application/json
      description: Get all plans
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/plans.Plan'
            type: array
        "400":
          description: Bad Request
          schema: {}
      summary: Get all plans
      tags:
      - plans
    post:
      consumes:
      - application/json
      description: Create a plan
      parameters:
      - description: Create Plan
        in: body
        name: plan
        required: true
        schema:
          $ref: '#/definitions/plans.CreatePlanDto'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/plans.Plan'
        "400":
          description: Bad Request
          schema: {}
      summary: Create a plan
      tags:
      - plans
  /plans/{id}:
    delete:
      consumes:
      - application/json
      description: Delete a plan
      parameters:
      - description: Plan ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema: {}
      summary: Delete a plan
      tags:
      - plans
    get:
      consumes:
      - application/json
      description: Get a plan
      parameters:
      - description: Plan ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/plans.Plan'
        "400":
          description: Bad Request
          schema: {}
      summary: Get a plan
      tags:
      - plans
    put:
      consumes:
      - application/json
      description: Update a plan
      parameters:
      - description: Plan ID
        in: path
        name: id
        required: true
        type: string
      - description: Update Plan
        in: body
        name: plan
        required: true
        schema:
          $ref: '#/definitions/plans.UpdatePlanDto'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/plans.Plan'
        "400":
          description: Bad Request
          schema: {}
      summary: Update a plan
      tags:
      - plans
  /plans/{user_id}/{day}:
    put:
      consumes:
      - application/json
      description: Update a plan by user and day
      parameters:
      - description: User ID
        in: path
        name: user_id
        required: true
        type: string
      - description: Day of week
        in: path
        name: day
        required: true
        type: string
      - description: Update Plan
        in: body
        name: plan
        required: true
        schema:
          $ref: '#/definitions/plans.UpdatePlanDto'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/plans.Plan'
        "400":
          description: Bad Request
          schema: {}
      summary: Update a plan by user and day
      tags:
      - plans
  /plans/user/{user_id}:
    get:
      consumes:
      - application/json
      description: Get a plan by user
      parameters:
      - description: User ID
        in: path
        name: user_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/plans.Plan'
            type: array
        "400":
          description: Bad Request
          schema: {}
      summary: Get a plan by user
      tags:
      - plans
  /plans/user/{user_id}/{day}:
    delete:
      consumes:
      - application/json
      description: Delete a plan by user and day
      parameters:
      - description: User ID
        in: path
        name: user_id
        required: true
        type: string
      - description: Day of week
        in: path
        name: day
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema: {}
      summary: Delete a plan by user and day
      tags:
      - plans
    get:
      consumes:
      - application/json
      description: Get a plan by user and day
      parameters:
      - description: User ID
        in: path
        name: user_id
        required: true
        type: string
      - description: Day of week
        in: path
        name: day
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/plans.Plan'
        "400":
          description: Bad Request
          schema: {}
      summary: Get a plan by user and day
      tags:
      - plans
...
schemes:
- http
- https
swagger: "2.0"